Updated CompressedData
I put some stuff in CompressedData from what I found on how to compress in java. This compresses it in a zip file, so if we want something else, I can try to find something else.

Also I just copy-pasted this and it might be completely wrong, so change stuff if needed.